Move attachment picker to molecules
